==> notebooks/2_splitting_data.ipynb <==
## Expected Goals (xG) Model - Data Splitting Pipeline

This notebook is used to split the imported data into competition-specific csv files.<br>
Feel free to work with the events file directly if you want, this step isn't mandatory.<br>
<i>**Note**</i> : MLS appears in this notebooks because this is an older version.

from pyspark.sql import SparkSession

spark = SparkSession.builder.appName("Splitting_Data").getOrCreate()

### 1. Splitting matches data

matches_df = spark.read.csv("../data/raw/matches.csv", header=True, inferSchema=True)

dataframes = {}
for competition in matches_df.groupBy('competition').count().orderBy('count', ascending=False).distinct().collect():
    dataframes[competition[0]] = matches_df.filter(matches_df.competition == competition[0])

for competition, df in dataframes.items():
    print(competition, ':', df.count())

for competition, df in dataframes.items():
    df.toPandas().to_csv(f"../data/split_data/matches/{competition}.csv", index=False)
    print(f"{competition} has been written to CSV.")

### 2. Splitting events data

events_df = spark.read.csv("../data/raw/events.csv", header=True, inferSchema=True, sep=";")

events_df.groupBy('competition').count().orderBy('count', ascending=False).show(truncate=False)

dataframes = {}
for competition in events_df.groupBy('competition').count().orderBy('count', ascending=False).distinct().collect():
    dataframes[competition[0]] = events_df.filter(events_df.competition == competition[0])

for competition, df in dataframes.items():
    df.coalesce(1).write.csv(f"../data/split_data/events/{competition}.csv", header=True, sep=";")
    print(f"{competition} events have been written to CSV.")

### 3. Splitting frames data

matches_df = spark.read.csv("../data/raw/matches.csv", header=True, inferSchema=True)

frames_df = spark.read.csv("../data/raw/frames.csv",header=True, inferSchema=True)

df_frames = {}
for comp in matches_df.groupBy('competition').count().distinct().collect():
    print(comp[0])
    df_ids = []
    for matchId in matches_df.filter(matches_df.competition == comp[0]).select('match_id').collect():
        df_ids.append(matchId[0])
    df_frames[comp[0]] = frames_df.filter(frames_df.match_id.isin(df_ids))

for comp,df in df_frames.items():
    print(comp, ':', df.count())

for comp, df in df_frames.items():
    if df.count() > 0:
        df.coalesce(1).write.csv(f"../data/split_data/frames/{comp}.csv", header=True)
        print(f"{comp} frames have been written to CSV.")

### 4. Splitting lineups

matches_df = spark.read.csv("../data/raw/matches.csv", header=True, inferSchema=True)

lineups_df = spark.read.csv("../data/raw/lineups.csv", header=True, inferSchema=True)

df_lineups = {}
for comp in matches_df.groupBy('competition').count().distinct().collect():
    print(comp[0])
    df_ids = []
    for matchId in matches_df.filter(matches_df.competition == comp[0]).select('match_id').collect():
        df_ids.append(matchId[0])
    df_lineups[comp[0]] = lineups_df.filter(lineups_df.match_id.isin(df_ids))

for comp, df in df_lineups.items():
    print(comp, ':', df.count())

for comp, df in df_lineups.items():
    df.coalesce(1).write.csv(f"../data/split_data/lineups/{comp}.csv", header=True)
    print(f"{comp} lineups have been written to CSV.")
